Separating API code from domain code
