+++ api/routes/download.py
@@ -1,10 +1,28 @@
-import io
+"""
+
+Route for MMIF downloads.
+
+Example request:
+
+curl -X POST 127.0.0.1:8001/storeapi/download \
+    -H 'Content-Type: "application/json"' \
+    -d '{"workflow": {"swt-detection/v7.4": {"pretty": "true"}}}'
+
+curl -X POST 127.0.0.1:8001/storeapi/download \
+    -H 'Content-Type: "application/json"' \
+    -d '{"workflow": {"swt-detection/v6.1": {"useStitcher": "false", "runningTime": "true", "hwFetch": "true"}}}'
+
+curl -X POST 127.0.0.1:8001/storeapi/download \
+    -H 'Content-Type: "application/json"' \
+    --output tmp.zip \
+    -d '{"workflow": {"swt-detection/v6.1": {"useStitcher": "false", "runningTime": "true", "hwFetch": "true"}}, "guid": ["cpb-aacip-259-wh2dcb8p","cpb-aacip-c72fd5cbadc"]}'
+
+"""
+
 import json
 import os
-import zipfile as zf
 from pathlib import Path
 from typing import Union
-from zipfile import ZIP_DEFLATED
 
 from flask import request, jsonify, Blueprint, send_file
 
@@ -12,12 +30,12 @@ from mmif.utils.workflow_helper import generate_param_hash
 from mmif import utils, Mmif
 
 from api import STORAGE_DIR
+from api.model.storage import get_mmif_for_guid, create_zipfile
 from api.errors import StorageServerError
-from api.utils import hash_from_dictionary
 
 
 bp = Blueprint('mmif_download', __name__)
-print(f'{bp} import_name={bp.import_name} __name__={__name__}')
+#print(f'{bp} import_name={bp.import_name} __name__={__name__}')
 
 
 API_PREFIX = '/storeapi'
@@ -42,16 +60,19 @@ def download_mmif():
 
     """
     data = json.loads(request.data.decode('utf-8'))
+    print('>>>', data)
     # build workflow storage path from request JSON
     # e.g. {"workflow": {"swt-detection/v2.0": {"pretty": "True"}}}
     # becomes "swt-detection/v2.0/paramhash"
     wfid_segments = []
     for clams_app, params in data["workflow"].items():
+        print('>>>', clams_app, params)
         wfid_segments.append(clams_app)
         try:
             wfid_segments.append(generate_param_hash(params))
         except AttributeError:
             wfid_segments.append(generate_param_hash({}))
+        print('>>>', wfid_segments)
     wfid = '/'.join(wfid_segments)
     # get number of views for rewind if necessary
     num_views = len(data.get('workflow', []))
@@ -77,7 +98,7 @@ def zero_guid_download_response(workflow_id: Union[str, Path]):
     and the list of files found there. This allows clients to utilize the api without
     downloading files (for working with local files).
     """
-    filenames = [p.stem for p in Path(workflow_id).glob('*')]
+    filenames = [p.stem for p in Path(workflow_id).glob('*.mmif')]
     return jsonify({'workflow': workflow_id, 'filenames': filenames})
 
 
@@ -95,72 +116,13 @@ def single_guid_download_response(workflow_id: str, guid: str, num_views: int):
 
 def multi_guid_download_response(workflow_id: str, guids: list, num_views: int):
     """
-    When retrieving multiple MMIFs for a workflow, we construct a json object to
-    store each guid as a key and each MMIF as the value.
+    When retrieving multiple MMIFs for a workflow, we return a zip file.
     """
-    errors = dict()
-    mem_file = io.BytesIO()
-    with zf.ZipFile(mem_file, 'w', ZIP_DEFLATED) as mmif_zip:
-        for guid in guids:
-            try:
-                # mmif = get_mmif_for_guid(workflow, guid, num_views)
-                # instead of using get_mmif_for_guid and needing to re-dump mmif
-                mmif_name = guid + ".mmif"
-                path = os.path.join(workflow_id, mmif_name)
-                mmif_zip.write(filename=path, arcname=f'multi-guid-response/files/{mmif_name}')
-            except FileNotFoundError:
-                errors[guid] = {"Error": f"Did not find {guid}"}
-        mmif_zip.writestr(zinfo_or_arcname="multi-guid-response/workflow_path.txt", data=workflow_id)
-        error_dump = json.dumps(errors, indent=2)
-        mmif_zip.writestr(zinfo_or_arcname="multi-guid-response/errors.json", data=error_dump)
-    mem_file.seek(0)
     # User will need to add '--output <FILE>' arg to curl request
     # NOTE (mv 12/12/25), the --output is needed even with the use of download_name
-    # below. In fact, it still works for me withoutremove that parameter, but keeping
-    # it anyway.
+    # below. In fact, it still works for me without removing that parameter, but
+    # keeping it anyway.
+    mem_file = create_zipfile(workflow_id, guids)
     return send_file(
         mem_file, mimetype='zip', as_attachment=True,
         download_name='multi-guid-response.zip')
-
-
-def get_mmif_for_guid(workflow_id: str, guid: str, num_views: int):
-    """
-    Retrieve the MMIF file for a workflow and GUID. If none was found raise a
-    StorageServerError.
-    """
-    guid = guid + ".mmif"
-    path = os.path.join(workflow_id, guid)
-    # if filepath exists, we can return it
-    try:
-        with open(path, 'r') as file:
-            mmif = json.loads(file.read())
-        return mmif
-    # otherwise we will use the rewinder to check if the user provided a prefix of a
-    # mmif workflow that we have previously stored
-    except FileNotFoundError:
-        try:
-            return rewind_time(workflow_id, guid, num_views)
-        except FileNotFoundError:
-            # the rewinder does not always succeed so we catch this exception again
-            # and raise an application-specific exception
-            raise StorageServerError(f'Did not find: {guid.split(".")[0]}')
-
-
-def rewind_time(workflow_id, guid, num_views):
-    """
-    This method takes in a workflow (path), a guid, and a number of views, and uses
-    os.walk to iterate through directories that begin with that workflow. It takes
-    the first mmif file that matches the guid and uses the rewind feature to include
-    only the views indicated by the workflow.
-    """
-    for home, dirs, files in os.walk(workflow_id):
-        # find mmif with matching guid to rewind
-        for file in files:
-            if guid == file:
-                # rewind the mmif
-                with open(os.path.join(home, file), 'r') as f:
-                    mmif = Mmif(f.read())
-                    # we need to calculate the number of views to rewind
-                    rewound = utils.rewind.rewind_mmif(mmif, len(mmif.views) - num_views)
-                return rewound.serialize()
-    raise FileNotFoundError
