+++ pyproject.toml
@@ -13,5 +13,9 @@ dependencies = [
 ]
 
 [tool.uv.sources]
-#mmif-python = { path = "packages/mmif_python-1.2.2.tar.gz" }
 inspector-mv = { path = "packages/inspector_mv-0.0.3.tar.gz" }
+
+[dependency-groups]
+dev = [
+    "pytest>=9.0.2",
+]
